Pkg/steps: Dont compile testhelpers into production binary
